ramips: ethernet: ralink: do a reset of the ethernet & switch together
The ethernet and switch needs to have the reset at the same time to
ensure a DMA panic doesn't happen.
The DMA panic seems to only happen on the rt503x devices as the SoCs
can have different versions on the ethernet, switch and DMA cores.

Do a partial revert of the reset api change and ignore the reset controller.

Tested-on: zyxel_nbg-419n
Fixes: 8569bc5e0d1b ("ramips: ethernet: ralink: rewrite esw_rt3050 to support link states")
Fixes: 60fadae62b64 ("ramips: ethernet: ralink: move reset of the esw into the esw instead of fe")
Fixes: 694561ae601a ("ramips: ethernet: ralink: use the reset controller api for esw & ephy")
Signed-off-by: Alexander Couzens <lynxis@fe80.eu>